+++ cleo_plugins/Audio/CSoundSystem.cpp
@@ -4,16 +4,18 @@
 #include "CLEO_Utils.h"
 #include "CAEAudioHardware.h"
 #include "CCamera.h"
+#include "CPad.h"
+#include "CVector.h"
 
 namespace CLEO
 {
     bool CSoundSystem::useFloatAudio = false;
     bool CSoundSystem::allowNetworkSources = true;
-    BASS_3DVECTOR CSoundSystem::pos(0.0, 0.0, 0.0);
-    BASS_3DVECTOR CSoundSystem::vel(0.0, 0.0, 0.0);
-    BASS_3DVECTOR CSoundSystem::front(0.0, -1.0, 0.0);
-    BASS_3DVECTOR CSoundSystem::top(0.0, 0.0, 1.0);
     eStreamType CSoundSystem::LegacyModeDefaultStreamType = eStreamType::None;
+    CVector CSoundSystem::position(0.0, 0.0, 0.0);
+    CVector CSoundSystem::velocity(0.0, 0.0, 0.0);
+    bool CSoundSystem::skipFrame = true;
+    float CSoundSystem::timeStep = 0.02f;
     float CSoundSystem::masterSpeed = 1.0f;
     float CSoundSystem::masterVolumeSfx = 1.0f;
     float CSoundSystem::masterVolumeMusic = 1.0f;
@@ -102,8 +104,7 @@ namespace CLEO
         }
 
         if (BASS_Init(deviceIndex, 44100, BASS_DEVICE_3D, RsGlobal.ps->window, nullptr) &&
-            BASS_Set3DFactors(1.0f, 3.0f, 80.0f) &&
-            BASS_Set3DPosition(&pos, &vel, &front, &top))
+            BASS_Set3DFactors(1.0f, 3.0f, 1.0f))
         {
             TRACE("SoundSystem initialized");
 
@@ -126,7 +127,6 @@ namespace CLEO
             }
 
             initialized = true;
-            BASS_Apply3D();
             return true;
         }
 
@@ -202,50 +202,49 @@ namespace CLEO
         {
             if (paused) Resume();
 
-            // get game globals
+            // update globals            
+            timeStep = 0.001f * (CTimer::m_snTimeInMillisecondsNonClipped - CTimer::m_snPreviousTimeInMillisecondsNonClipped); // delta in seconds
             masterSpeed = CTimer::ms_fTimeScale;
             masterVolumeSfx = AEAudioHardware.m_fEffectMasterScalingFactor * 0.5f; // fit to game's sfx volume
             masterVolumeMusic = AEAudioHardware.m_fMusicMasterScalingFactor * 0.5f;
 
-            // camera movements
-            CMatrixLink * pMatrix = nullptr;
-            CVector * pVec = nullptr;
-            if (TheCamera.m_matrix)
+            // prevent camera jump-cut glitches
+            int skipFramePrev = skipFrame;
+            skipFrame = TheCamera.m_bJust_Switched || TheCamera.m_bCameraJustRestored || CPad::GetPad(0)->JustOutOfFrontEnd;
+
+            CVector prevPos = position;
+            position = TheCamera.GetPosition(); // get new
+
+            // new camera velocity
+            if (!skipFrame)
             {
-                pMatrix = TheCamera.m_matrix;
-                pVec = &pMatrix->pos;
-            }
-            else pVec = &TheCamera.m_placement.m_vPosn;
+                CVector vel = position - prevPos;
+                vel /= timeStep; // meters per second
 
-            BASS_3DVECTOR prevPos = pos;
-            pos = BASS_3DVECTOR(pVec->y, pVec->z, pVec->x);
+                if (!skipFramePrev)
+                {
+                    // averaging to smooth artifacts caused by GTA's janky mouse camera control
+                    velocity = (velocity * 2.0f) + vel;
+                    velocity /= 3.0f;
+                }
+                else
+                {
+                    velocity = vel;
+                }
 
-            // calculate velocity
-            vel = prevPos;
-            vel.x -= pos.x;
-            vel.y -= pos.y;
-            vel.z -= pos.z;
-            auto timeDelta = 0.001f * (CTimer::m_snTimeInMillisecondsNonClipped - CTimer::m_snPreviousTimeInMillisecondsNonClipped);
-            vel.x *= timeDelta;
-            vel.y *= timeDelta;
-            vel.z *= timeDelta;
-
-            // setup the ears
-            if (!TheCamera.m_bJust_Switched && !TheCamera.m_bCameraJustRestored) // avoid camera change/jump cut velocity glitches
-            {
                 BASS_Set3DPosition(
-                    &pos,
-                    &vel,
-                    pMatrix ? &BASS_3DVECTOR(pMatrix->at.y, pMatrix->at.z, pMatrix->at.x) : nullptr,
-                    pMatrix ? &BASS_3DVECTOR(pMatrix->up.y, pMatrix->up.z, pMatrix->up.x) : nullptr
+                    &toBass(position),
+                    &toBass(velocity),
+                    &toBass(TheCamera.GetForward()),
+                    &toBass(TheCamera.GetUp())
                 );
             }
 
             // process streams
-            for(auto stream : streams) stream->Process();
+            for (auto stream : streams) stream->Process();
 
-            // apply above changes
-            BASS_Apply3D();
+            // apply changes
+            if (!skipFrame) BASS_Apply3D();
         }
     }
 }